more permissions and better views for them
**CRITICAL NOTE**

If you have previously checked out PLAS, there are a few permissions I altered by hand in older migrations. These permissions are not currently in use and won't likely be in use for a month or so.

If these permissions are not manually modified, you will merely have to add them in the permissions tool.

To manually modify them, open the database with `sqlite3 db/development.db` and run this:

    update permissions set code = "tournaments.record" where code = "tournament.record";
    update permissions set code = "tournaments.create" where code = "tournament.create";
    update permissions set code = "tournaments.start" where code = "tournament.start";
    update permissions set code = "tournaments.change_team" where code = "tournament.change_team";
    update permissions set code = "events.start" where code = "event.start"